+++ src/main/java/com/example/KeyboardArenaProject/controller/freeBoard/FreeBoardController.java
@@ -149,41 +149,6 @@ public class FreeBoardController {
     }
 
     //좋아요
-    @ResponseBody
-    @PostMapping("/api/like")
-    public void like(@RequestParam String boardId,@RequestParam String id){
-        UserBoardCompositeKey userBoardCompositeKey = UserBoardCompositeKey.builder().id(id).boardId(boardId).build();
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if(!authentication.getPrincipal().equals("anonymousUser")) {
-            int likes;
-            if(likeService.findById(userBoardCompositeKey)!=null){
-                if(!likeService.findById(userBoardCompositeKey).isIfLike()){
-                    likes = likeService.clickLikeOne(userBoardCompositeKey);
 
-                }else{
-                    likes = likeService.clickLikeTwo(userBoardCompositeKey);
-                }
-
-            }else{
-                likes = likeService.save(userBoardCompositeKey);
-            }
-            freeBoardService.updateBoardLikes(likes, userBoardCompositeKey.getBoardId());
-        }
-    }
-
-    //댓글
-    @PostMapping("/comments/{board_id}")
-    public String saveComment(@PathVariable String board_id,@RequestBody Map<String,String> content){
-        Comment comment = new Comment(board_id,content.get("content"), userService.getCurrentUserInfo().getId(),userService.getCurrentUserInfo().getNickname());
-        commentService.saveComment(comment);
-        return "redirect:/board/"+board_id;
-    }
-
-    @DeleteMapping("/comments/{comment_id}")
-    public String deleteComment(@PathVariable String comment_id){
-        String boardId = commentService.findCommentById(comment_id).getBoardId();
-        commentService.deleteComment(comment_id);
-        return "redirect:/board/"+boardId;
-    }
 
 }
